A working rbot plugin for asking about assigned tickets.
  !tickets
            will show the tickets assigned for the user in redmine
            that have the same login name at redmine as the nick
            in IRC.

  !tickets user
             will show the tickets assigned for the user in redmine
             that have the provided login name.

Default setting is to show five assigned tickets. This can be adjusted
via the rbot config system.

A redmine administrator user is needed, add username and password to the
rbot config system.

A redmine instance URI scheme must be added, default is
http://localhost.

A list of channels where the plugin is applicable is needed, default
channel is #kaos as that's what we use =)

See !config list redmine